@@ -40,15 +40,19 @@
   "author": "Colin Aulds",
   "license": "ISC",
   "dependencies": {
+    "@types/react": "^0.14.55",
+    "@types/react-dom": "^0.14.20",
     "body-parser": "^1.15.2",
     "express": "^4.14.0",
     "isomorphic-fetch": "^2.2.1",
     "jquery": "^3.1.1",
+    "mongoose": "^4.7.5",
     "oauthsimple": "0.0.1",
     "passport": "^0.3.2",
-    "react": "^15.3.2",
+    "passport-http": "^0.3.0",
+    "react": "^15.4.1",
     "react-bootstrap": "^0.30.7",
-    "react-dom": "^15.3.2",
+    "react-dom": "^15.4.1",
     "react-places-autocomplete": "^2.3.2",
     "react-redux": "^4.4.5",
     "react-router": "^2.8.1",
@@ -57,6 +61,7 @@
     "redux-thunk": "^2.1.0"
   },
   "devDependencies": {
+    "awesome-typescript-loader": "^3.0.0-beta.17",
     "babel-core": "^6.14.0",
     "babel-loader": "^6.2.5",
     "babel-plugin-transform-object-rest-spread": "^6.20.2",
@@ -70,7 +75,9 @@
     "less-plugin-clean-css": "^1.5.1",
     "mocha": "^3.0.2",
     "react-addons-test-utils": "^15.3.2",
+    "source-map-loader": "^0.1.5",
     "supervisor": "^0.11.0",
+    "typescript": "^2.1.4",
     "uglify-js": "^2.7.3",
     "webpack": "^1.13.2"
   }
