@@ -1,4 +1,4 @@
-import {POST_SURVEY} from "./actions";
+import {POST_SURVEY, POST_USER, GET_USER, GET_SURVEY} from "./actions";
 import { handle } from 'redux-pack';
 
 var initialState = {
@@ -21,8 +21,43 @@ var SurpayApp = function(state, action) {
         failure: s => ({ ...s, surveyError: payload }),
         success: s => ({ ...s, survey: payload }),
       });
+      case POST_USER:
+        return handle(state, action, {
+          start: s => ({
+            ...s,
+            isLoading: true,
+            surveyError: null
+          }),
+          finish: s => ({ ...s, isLoading: false }),
+          failure: s => ({ ...s, userError: payload }),
+          success: s => ({ ...s, user: payload }),
+        });
+      case GET_USER:
+        console.log(action);
+        return handle(state, action, {
+          start: s => ({
+            ...s,
+            isLoading: true,
+            surveyError: null
+          }),
+          finish: s => ({ ...s, isLoading: false }),
+          failure: s => ({ ...s, userError: payload }),
+          success: s => ({ ...s, user: payload }),
+        });
+      case GET_SURVEY:
+        console.log(action);
+        return handle(state, action, {
+          start: s => ({
+            ...s,
+            isLoading: true,
+            surveyError: null
+          }),
+          finish: s => ({ ...s, isLoading: false }),
+          failure: s => ({ ...s, userError: payload }),
+          success: s => ({ ...s, user: payload }),
+        });
     default:
-      return state;
+        return state;
   }
 };
 
